0969 refactor
