+++ modules/burials/lib/burials/version.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Burials
+  # version
+  VERSION = '0.1.0'
+end
