+++ modules/pensions/lib/pensions/version.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Pensions
+  # version
+  VERSION = '0.1.0'
+end
